% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCoords.R
\name{getCoords}
\alias{getCoords}
\title{Get or create coordinates from Data structure or}
\usage{
getCoords(data)
}
\arguments{
\item{data}{Data structure to check, could be from:
(1) Data structure from geomorph or created by readTem()
(2) Data structure with a 3D array named coords
(3) Data structure with a 3D array named aligned$coords
(3) 3D array of coordinate values}
}
\value{
A 3D array of coordinate values
}
\description{
Get or create coordinates from Data structure or
}
\examples{
faces$coords <- getCoords(faces)
}

