% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Create a data frame with data from data$info and PCs from data$pca}
\usage{
getData(data, pcs = c())
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}

\item{pcs}{An array of PCs to add (Default NULL = include all)}
}
\value{
A data frame
}
\description{
Create a data frame with data from data$info and PCs from data$pca
}
\examples{
mydata <- saveData(data, pcs = rep(1:10), save.name = "~/Desktop/mydata.csv")
}

