% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Get PCA Data}
\usage{
getPCA(data)
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}
}
\value{
A table summarizing the percent variation explained by each PC and the set of PC scores
}
\description{
Get PCA Data
}
\examples{
faces$pca <- getPCA(faces)
}
