% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gifImages.R
\name{gifImages}
\alias{gifImages}
\title{Create images for a moving gif}
\usage{
gifImages(data, pc = 1, dir = "", start = -1, end = 1, steps = 20,
  height = 400, width = 300)
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}

\item{pc}{The PC to manipulate (Defaults to 1)}

\item{dir}{The directory to save images in (Defaults to working directory)}

\item{start}{The starting SD (Defaults to -1)}

\item{end}{The ending SD (Defaults to 1)}

\item{steps}{The number of steps in the continuum (Defaults to 20)}

\item{height}{The height of the output png in pixels (Defaults to 400)}

\item{width}{The width of the output png in pixels (Defaults to 400)}
}
\description{
Create images for a moving gif
}
\examples{
gifImages(faces, pc = 1, dir = "~/Desktop/sd1")
}
