% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphCompare2.R
\name{graphCompare2}
\alias{graphCompare2}
\title{Create a graph plotting two PCs and visualising min and max for those PCs.}
\usage{
graphCompare2(data, axis1 = 1, axis2 = 2, gp = c(),
  col.gp = c("#d11141", "#f47835", "#ffc425", "#00b159", "#00aedb",
  "#a200ff"))
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}

\item{axis1}{The PC to be shown on the x axis (Defaults to 1)}

\item{axis2}{The PC to be shown on the y axis (Defaults to 2)}

\item{gp}{An array of group names for each shape (Defaults to NULL array: c())}

\item{col.gp}{An array of colours for the groups (Defaults to ROYGBP: c("#d11141", "#f47835", "#ffc425", "#00b159", "#00aedb", "#a200ff"))}
}
\value{
TRUE
}
\description{
Create a graph plotting two PCs and visualising min and max for those PCs.
}
\examples{
faces$pca <- graphCompare2(faces$aligned$coords, 2, 3)
}
