% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphMinMax.R
\name{graphMinMax}
\alias{graphMinMax}
\title{Create a graph comparing min and max values on one PC}
\usage{
graphMinMax(data, pc = 1, mag = 1, method = "TPS")
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}

\item{pc}{The PC to be visualised (Defaults to 1)}

\item{mag}{The magnification factor of the visualised difference (Deafults to 1)}

\item{method}{The graphing method for plotRefToTarget (Defaults to "TPS")}
}
\description{
Create a graph comparing min and max values on one PC
}
\examples{
faces$pca <- graphMinMax(faces, 2)
}
