% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphShape.R
\name{graphShape}
\alias{graphShape}
\title{Project (and optionally graph and/or save to a file) a shape using an array of SDs for each PC}
\usage{
graphShape(data, set.pc = c(0), by = "SD", show.image = F,
  save.name = "")
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}

\item{set.pc}{An array of values or SDs for each PC (unspecified PCs will be 0) (Defaults to c(0))}

\item{by}{Whether set.pc specifies "SD" or "value" (Defaults to "SD")}

\item{show.image}{Whether to display a graph of the resulting array (Defaults to F)}

\item{save.name}{A file name to save to (Defaults to ""; no saving)}
}
\value{
An array of coordinate values for the resulting shape
}
\description{
Project (and optionally graph and/or save to a file) a shape using an array of SDs for each PC
}
\examples{
face.sd3.min <- graphShape(faces, c(0,0,-2), show.image=T)
face.sd3.max <- graphShape(faces, c(0,0,2), show.image = F, savename="sd3max.tem")
}

