% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/groupByName.R
\name{groupByName}
\alias{groupByName}
\title{Allocate images to groups based on filenames and regex patterns}
\usage{
groupByName(img.list, names, patterns)
}
\arguments{
\item{img.list}{A list of filenames}

\item{names}{A list of the groups to be allocated}

\item{patterns}{A list of regex patterns to match for group allocation}
}
\value{
A list of group names for each filename in img.list
}
\description{
Allocate images to groups based on filenames and regex patterns
}
\examples{
faces$info$group <- groupByName(faces$info$filename, names=c("male","female"), patterns=c("^male/*","^female/*")))
}

