% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustesAlign.R
\name{procrustesAlign}
\alias{procrustesAlign}
\title{Procrustes align frlgmm object.}
\usage{
procrustesAlign(data)
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}
}
\value{
3D Array of procrustes-aligned coordinates
}
\description{
Uses geomorph::gpagen with FRL defaults.
Proj = TRUE: project template landmark data into tangent space.
ProcD = FALSE: Does not rotate output to principal axes, as this usually turns heads sideways.
PrinAxes = FALSE: Uses minimising bending energy to slide semilandmarks if curvepts defined.
}
\examples{
faces$aligned <- procrustesAlign(faces)
}
