% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readTems.R
\name{readTems}
\alias{readTems}
\title{Read Psychomorph template files from a directory or list of directories.}
\usage{
readTems(tem.loc, semi.landmarks = 0, curvefile = "")
}
\arguments{
\item{tem.loc}{Directory or csv table of filenames where .tem files are found}

\item{semi.landmarks}{Number of curve points to graphically define}

\item{curvefile}{Name of curvepts file}
}
\value{
A data structure ready for geomorph, with land, curvepts, and file info
}
\description{
Read Psychomorph template files from a directory or list of directories.
}
\examples{
faces <- readTems("~/Dropbox/faces")
faces <- readTems("~/Dropbox/faces/faceinfo.csv", curvefile = "~/Dropbox/faces/curveslide.csv"))
}

