% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTem.R
\name{saveTem}
\alias{saveTem}
\title{Save coordinates to a file}
\usage{
saveTem(coords, save.name = "", sample.tem = "", align.to = c(1, 2),
  data = NULL)
}
\arguments{
\item{coords}{An array of projected coordinates}

\item{save.name}{A file name to save to (Defaults to ""; no saving)}

\item{sample.tem}{An optional array or filename for a sample tem file from which to get lines}

\item{align.to}{An array of coordinates numbers to align the tem to (taken from the sample tem)}

\item{data}{An frlgmm data object to align to}
}
\description{
Save coordinates to a file
}
\examples{
saveTem(coords, save.name="sd3max.tem", sample.tem=faces$info$filename[1]))
}

