% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selectPCs.R
\name{selectPCs}
\alias{selectPCs}
\title{Select PCs using different criteria}
\usage{
selectPCs(data, method = "mean eigenvalue")
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}

\item{method}{The method to use to choose PCs (Default "mean eigenvalue")}
}
\value{
A list of chosen PCs
}
\description{
Select PCs using different criteria
}
\examples{
chosen.pcs <- selectPCs(data)
}

