% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uprightFace.R
\name{uprightFace}
\alias{uprightFace}
\title{Rotate aligned data to upright using eye point locations}
\usage{
uprightFace(coords, left.eye = 1, right.eye = 2)
}
\arguments{
\item{coords}{A 3D Array of procrustes-aligned coordinates}

\item{left.eye}{Index of the left eye point (Defaults to 1)}

\item{right.eye}{Index of the right eye point (Defaults to 2)}
}
\value{
A 3D Array of procrustes-aligned coordinates
}
\description{
Rotate aligned data to upright using eye point locations
}
\examples{
faces$aligned$coords <- uprightFace(faces$aligned$coords)
}

