% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustesAlign.R
\name{procrustesAlign}
\alias{procrustesAlign}
\title{Procrustes align frlgmm object with FRL defaults.}
\usage{
procrustesAlign(data)
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}
}
\value{
3D Array of procrustes-aligned coordinates
}
\description{
Procrustes align frlgmm object with FRL defaults.
}
\examples{
faces$aligned <- procrustesAlign(faces)
}

