% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTems.R
\name{readTems}
\alias{readTems}
\title{Read Psychomorph template files}
\usage{
readTems(tem.loc, semi.landmarks = 0, curvefile = "", analyse = TRUE,
  method = "broken-stick")
}
\arguments{
\item{tem.loc}{Directory or csv table of filenames where .tem files are found}

\item{semi.landmarks}{Number of curve points to graphically define}

\item{curvefile}{Name of curvepts file}

\item{analyse}{Whether to include aligned data and PCA analysis using FRL defaults (Defaults to TRUE)}

\item{method}{The method to use to choose PCs (Default "broken-stick")}
}
\value{
A data structure with land, curvepts, info and (optionally) aligned and PCA analysed
}
\description{
\code{readTems} reads Psychomorph/WebMorph .tem files from a directory, a list of directories,
or a CSV table with the filenames of each template file to be loaded (e.g., "subdir/file1.tem")
}
\examples{
faces <- readTems("~/Dropbox/faces")
faces <- readTems("~/Dropbox/faces/faceinfo.csv", curvefile = "~/Dropbox/faces/curveslide.csv"))
}

