% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readTems.R
\name{readTems}
\alias{readTems}
\title{Read Psychomorph template files from a directory or list of directories.}
\usage{
readTems(tem.loc, semi.landmarks = 0, curvefile = "", analyse = TRUE)
}
\arguments{
\item{tem.loc}{Directory or csv table of filenames where .tem files are found}

\item{semi.landmarks}{Number of curve points to graphically define}

\item{curvefile}{Name of curvepts file}

\item{analyse}{Whether to include aligned data and PCA analysis using FRL defaults (Defaults to TRUE)}
}
\value{
A data structure with land, curvepts, info and (optionally) aligned and PCA analysed
}
\description{
Read Psychomorph template files from a directory or list of directories.
}
\examples{
faces <- readTems("~/Dropbox/faces")
faces <- readTems("~/Dropbox/faces/faceinfo.csv", curvefile = "~/Dropbox/faces/curveslide.csv"))
}

