% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveTem.R
\name{saveTem}
\alias{saveTem}
\title{Save coordinates to a file}
\usage{
saveTem(coords, save.name = "", sample.tem = "", align.to = NULL)
}
\arguments{
\item{coords}{An array of projected coordinates}

\item{save.name}{A file name to save to (Defaults to ""; no saving)}

\item{sample.tem}{An optional filename for a sample tem file from wich to get lines}

\item{align.to}{An array of coordinates to align the tem to (usually the mean of the original set)}
}
\value{
true if file saved successfully
}
\description{
Save coordinates to a file
}
\examples{
saveTem(coords, save.name="sd3max.tem", sample.tem=faces$info$filename[1]))
}

