% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPCs.R
\name{selectPCs}
\alias{selectPCs}
\title{Select PCs using different criteria}
\usage{
selectPCs(data, method = "broken-stick", total.var = 0.95)
}
\arguments{
\item{data}{Data structure from geomorph or created by readTem()}

\item{method}{The method to use to choose PCs (Default "broken-stick")}

\item{total.var}{Total variance to choose for "total variance" method (deafult .95)}
}
\value{
A list of chosen PCs
}
\description{
\code{selectPCs} returns a list of significant PCs as chosen by one of 3 methods
(from D.A. Jackson (1993). Ecology, 74, 2204-2214). Defaults to the most accurate
and conservative "broken-stick" method. Other methods are "Kaiser-Guttman" (PCs
with eigenvalues greater than the mean eigenvalue) and "total variance" (PCs
explaining at least total.var variance)
}
\examples{
chosen.pcs <- selectPCs(data)
}

